Factorial used BigDecimal instead of long This will allow for MUCH larger numbers.  Added a unit tests to try a very larger factorial.

+++ src/main/java/org/fross/rpncalc/StackCommands.java
@@ -356,7 +356,7 @@ public class StackCommands {
 			Output.printColorln(Ansi.Color.CYAN, "Warning: The decimal portion of '" + calcStack.peek().toString() + "' has been dropped for the calculation");
 		}
 		
-		BigDecimal result = Math.Factorial(calcStack.pop().toBigInteger().longValue());
+		BigDecimal result = Math.Factorial(calcStack.pop());
 		calcStack.push(result);
 	}
